Modify Int -> Int16 for 'children' in ResponseAPI.

+++ GoloSwift/Models/ResponseAPI.swift
@@ -63,7 +63,7 @@ public struct ResponseAPIPost: Decodable {
     public let pending_payout_value: String
     public let total_payout_value: String
     public let reblogged_by: [String]?
-    public let children: Int
+    public let children: Int16
 
     // "2018-04-13T10:19:54"
     public let created: String
@@ -347,7 +347,7 @@ public struct ResponseAPIAllContentReply: Decodable {
     public let json_metadata: String
     public let url: String
     public let reblogged_by: [String]?
-    public let children: Int
+    public let children: Int16
 
     // "2018-04-13T10:19:54"
     public let created: String
